skyarclog/requirements.txt
# Core Python dependencies
typing>=3.7.4

# Database Listeners
mysql-connector-python>=8.0.26
psycopg2-binary>=2.9.1
pymongo>=3.12.0

# Cloud Logging
azure-monitor-opentelemetry>=1.0.0
boto3>=1.18.0
google-cloud-logging>=3.0.0

# OpenTelemetry (for Azure logging)
opentelemetry-api>=1.10.0
opentelemetry-sdk>=1.10.0

# Optional: Performance and async support
uvloop>=0.16.0  # Optional async support
